+++ lib/planet.rb
@@ -1,5 +1,3 @@
-require "pry"
-
 class Planet
   attr_reader :name, :color, :mass_kg, :distance_from_sun_km, :fun_facts
